[WIP] Implement SSL-Support for Server
Unfortunatly this does not seem to work like this. The unittest does not
work properly and it does also not work connecting with a seperate
script to the port.
The underlying technique should work though, as server.php and
testssl.php show. I think it has to do something with how
StreamSelectLoop handles the streams, but I have no idea what is wrong
with it.